+++ src/utils/constants.ts
@@ -2,3 +2,5 @@ export const OWNER = "ton-blockchain"
 export const REPO = "acton"
 
 export const BINARY_NAME = "acton"
+
+export const ACTON_TOML_FILE_NAME = "Acton.toml"